Fix for TypeError when running the NPM `build` script.
Due to a recent update to a babel react related library the `build:cjs`
and `build:esm` scripts fail with a TypeError related to jsx files
(pass.get(...) is not a function). To compensate We are tracking a
previously build `lib` folder into our fork. This may change in the
future when a known fix is found.